[Vansh] : added program to check whether students can vote or not

+++ gcr-codebase/control-flow/Level3/DayOfWeek.java
@@ -6,12 +6,12 @@ public class DayOfWeek {
         int d = Integer.parseInt(args[1]); // day
         int y = Integer.parseInt(args[2]); // year
         // Apply given formulas
-        int y0 = y - (14 - m) / 12;
-        int x = y0 + y0 / 4 - y0 / 100 + y0 / 400;
-        int m0 = m + 12 * ((14 - m) / 12) - 2;
-        int d0 = (d + x + (31 * m0) / 12) % 7;
+        double y0 = y - (14 - m) / 12;
+        double x = y0 + y0 / 4 - y0 / 100 + y0 / 400;
+        double m0 = m + 12 * ((14 - m) / 12) - 2;
+        double d0 = (d + x + 31 * m0 / 12) % 7;
 
         // Print day of the week
-        System.out.println("Day of the week is: " + d0);
+        System.out.println("Day of the week is: " + (int)d0);
     }
 }
